ddl: add dedicated materialized view drop actions
