+++ pkg/ddl/job_worker_test.go
@@ -224,6 +224,10 @@ func TestJobNeedGC(t *testing.T) {
 	require.True(t, ddl.JobNeedGC(job))
 	job = &model.Job{Type: model.ActionAddPrimaryKey, State: model.JobStateRollbackDone}
 	require.True(t, ddl.JobNeedGC(job))
+	job = &model.Job{Type: model.ActionDropMaterializedView, State: model.JobStateDone}
+	require.True(t, ddl.JobNeedGC(job))
+	job = &model.Job{Type: model.ActionDropMaterializedViewLog, State: model.JobStateDone}
+	require.True(t, ddl.JobNeedGC(job))
 	job = &model.Job{Type: model.ActionCreateMaterializedView, State: model.JobStateDone, TableID: 123}
 	require.False(t, ddl.JobNeedGC(job))
 	job = &model.Job{Type: model.ActionCreateMaterializedView, State: model.JobStateRollbackDone}
