+++ pkg/infoschema/BUILD.bazel
@@ -87,6 +87,7 @@ go_test(
     timeout = "short",
     srcs = [
         "bench_test.go",
+        "builder_mview_test.go",
         "builder_test.go",
         "infoschema_nokit_test.go",
         "infoschema_test.go",
@@ -99,7 +100,7 @@ go_test(
     ],
     embed = [":infoschema"],
     flaky = True,
-    shard_count = 40,
+    shard_count = 41,
     deps = [
         "//pkg/ddl/placement",
         "//pkg/domain",
